+++ src/kernel/trackers/alloc.rs
@@ -18,7 +18,7 @@ pub struct AllocEntry {
     // if there's data needed to be allocation before running program, then specify what content this is
     // If this value is none, then some sort of computation needs to be done before this appears.
     // also note that the size of the initial content is not the same as the size of the alloc itself!
-    pub initial_content: Option<Arc<Vec<f64>>>,
+    pub initial_content: Option<Arc<Vec<f32>>>,
 
     pub alloc_loc: Location,
     pub alloc_defined: bool,  // Alloc is already defined
